+++ modules/util/sync-engine/src/com/liferay/sync/engine/filesystem/SyncWatchEventProcessor.java
@@ -15,6 +15,8 @@
 package com.liferay.sync.engine.filesystem;
 
 import com.liferay.sync.engine.SyncEngine;
+import com.liferay.sync.engine.documentlibrary.util.BatchEvent;
+import com.liferay.sync.engine.documentlibrary.util.BatchEventManager;
 import com.liferay.sync.engine.documentlibrary.util.FileEventUtil;
 import com.liferay.sync.engine.model.SyncAccount;
 import com.liferay.sync.engine.model.SyncFile;
@@ -23,8 +25,6 @@ import com.liferay.sync.engine.model.SyncWatchEvent;
 import com.liferay.sync.engine.service.SyncAccountService;
 import com.liferay.sync.engine.service.SyncFileService;
 import com.liferay.sync.engine.service.SyncWatchEventService;
-import com.liferay.sync.engine.util.BatchEvent;
-import com.liferay.sync.engine.util.BatchEventUtil;
 import com.liferay.sync.engine.util.FileUtil;
 import com.liferay.sync.engine.util.SyncEngineUtil;
 
@@ -106,6 +106,10 @@ public class SyncWatchEventProcessor implements Runnable {
 		catch (Exception e) {
 			_logger.error(e.getMessage(), e);
 		}
+
+		BatchEvent batchEvent = BatchEventManager.getBatchEvent(_syncAccountId);
+
+		batchEvent.fireBatchEvent();
 	}
 
 	protected void addFile(SyncWatchEvent syncWatchEvent) throws Exception {
@@ -381,10 +385,6 @@ public class SyncWatchEventProcessor implements Runnable {
 	}
 
 	protected void doRun() throws Exception {
-		BatchEvent batchEvent = BatchEventUtil.getBatchEvent(_syncAccountId);
-
-		batchEvent.fireBatchEvent();
-
 		SyncWatchEvent lastSyncWatchEvent =
 			SyncWatchEventService.fetchLastSyncWatchEvent(_syncAccountId);
